Correcting logic defect regarding the value of an Ace.
Assume the Player has King/Ace, yet for some reason chooses to hit. The hit card is Ace. The previous algorithm would correctly compute a hand value of 12. If the player chose to hit again and this time received a 10 or Face Card, the algorithm would correctly assign a value of 22, however the ace-value flip entry criteria would also be met - sum is greater than 21 and has_ace is true, and would reset the total to 12 for a hand of KAAK.

Correction is to mark if the hand has a high-ace value and only perform the ace-value correction once by resetting the has-high-ace boolean once the ace-value is flipped to 1.